Grid Horizontal Coordinate system. Consolidate into concrete class in nc2.grid2. Add tests.
